refactor(ipc): share the daemon proxy between the window and a CLI
