+++ crates/tidemark/Cargo.toml
@@ -24,6 +24,7 @@ gtk = { version = "0.11.4", package = "gtk4", features = ["v4_22"] }
 # bundled Rubik available to GTK without installing it for the user.
 pango = { version = "0.22.3", features = ["v1_56"] }
 semver = "1.0.28"
+tidemark-ipc = { path = "../tidemark-ipc" }
 tidemark-types = { path = "../tidemark-types" }
 tracing = "0.1.44"
 tracing-subscriber = { version = "0.3.23", features = ["env-filter"] }
